split ClientServerDUnit for better balance in parallel test runs
